Various fixes and improvements
Reworked a lot of logic and added detection for a lot of edge case issues, like quote retweets not having the source tweet link element to base off of.

Also created a custom context menu so the file names when saving can be altered to always include the people's username and post status.