Responsiveness and layout improvements
Deferred the HQ image load until after adjusted layout is calculated to avoid the layout changes having to wait on the HQ file to load and thus being noticeable often.

Also improved the 4x4 panel layout to elongate further when the content inside allows it, instead of merely being a 1:1 square at most.

Side-By-Side layout also improved to not expand so much and instead try to show all the content.