Fix edited retweets not being processed/layout-adjusted
Reposts can have unique IDs from the original post, but these are not new entries in the response data, so it wasn't being tracked in my cache information. I now record the edited IDs for a post as well, so now these Retweet posts will be properly processed and not have that gap in their layout, nor be missing the right-click menu.

Also storing a bit less data in my cache now, might help a touch with performance on some devices.