Refactoring, performance improvements, HQ Video option, preroll AD block...
Did a bit of refactoring, and improved how some stuff is loaded, as well as ensuring no work is done to get higher quality content if the options are turned off, only grabbing it when you try to download it to your computer.

Added some handling of content in custom "CARDS", which are posts that source their content externally.
Related to this, people are able to mask the URL these cards open when clicked... which is a big security flaw. So I went ahead and removed the ability for that to happen when you click on the content. You must instead click on the link below the content. I also unmasked the link so you can see clearly that it's not simply a link to another profile....

Added blocking of video pre-roll ADs. They aren't really using them much yet, but I'm sure they will more as time goes on.

Added an option to toggle automatic loaded of the highest quality video source, since that can be heavy on performance and some people may only care for it to be highest when they click to Download it, which it will be.

Cleaned up some logic so behaviour is more consistent as well.